+++ examples/flag_example.rs
@@ -82,21 +82,62 @@ fn spawn_cloth(
     asset_server: Res<AssetServer>,
 ) {
     let flag_texture = asset_server.load("Bevy.png");
+    let material = materials.add(StandardMaterial {
+        base_color_texture: Some(flag_texture),
+        cull_mode: None,    // Option required to render back faces correctly
+        double_sided: true, // Option required to render back faces correctly
+        ..Default::default()
+    });
     let (size_x, size_y) = (30, 15);
     let mesh = rectangle_mesh((size_x, size_y), (Vec3::X * 0.5, -Vec3::Y * 0.5), Vec3::Z);
-    let cloth = ClothBuilder::new().with_fixed_points((0..size_y).map(|i| i * size_x));
+
+    // Regular Smooth Flag
+    let cloth = ClothBuilder::new().with_pinned_vertex_ids((0..size_y).map(|i| i * size_x));
     commands
         .spawn_bundle(PbrBundle {
-            mesh: meshes.add(mesh),
-            material: materials.add(StandardMaterial {
-                base_color_texture: Some(flag_texture),
-                cull_mode: None,    // Option required to render back faces correctly
-                double_sided: true, // Option required to render back faces correctly
-                ..Default::default()
-            }),
-            transform: Transform::from_xyz(0.0, 5.0, 0.0),
+            mesh: meshes.add(mesh.clone()),
+            material: material.clone(),
+            transform: Transform::from_xyz(0.0, 8.0, 5.0),
             ..Default::default()
         })
         .insert(cloth)
-        .insert(Name::new("Flag"));
+        .insert(Name::new("Regular Smooth Flag"));
+
+    // Regular Flat Flag
+    let cloth = ClothBuilder::new()
+        .with_pinned_vertex_ids((0..size_y).map(|i| i * size_x))
+        .with_flat_normals();
+    commands
+        .spawn_bundle(PbrBundle {
+            mesh: meshes.add(mesh.clone()),
+            material: material.clone(),
+            transform: Transform::from_xyz(0.0, 8.0, -5.0),
+            ..Default::default()
+        })
+        .insert(cloth)
+        .insert(Name::new("Regular Flat Flag"));
+
+    // TODO: enable when bevy releases vertex color support
+    // color flag
+    // let mut mesh = mesh;
+    // if let Some(VertexAttributeValues::Float32x3(positions)) =
+    //     mesh.attribute(Mesh::ATTRIBUTE_POSITION)
+    // {
+    //     let colors: Vec<[f32; 4]> = positions
+    //         .iter()
+    //         .map(|[r, g, b]| [(1. - *r) / 2., (1. - *g) / 2., (1. - *b) / 2., 1.])
+    //         .collect();
+    //     println!("{}", colors.len());
+    //     mesh.insert_attribute(Mesh::ATTRIBUTE_COLOR, colors);
+    // }
+    // let cloth = ClothBuilder::new().with_pinned_vertex_ids((0..size_y).map(|i| i * size_x));
+    // commands
+    //     .spawn_bundle(PbrBundle {
+    //         mesh: meshes.add(mesh),
+    //         material,
+    //         transform: Transform::from_xyz(0.0, 8.0, -5.0),
+    //         ..Default::default()
+    //     })
+    //     .insert(cloth)
+    //     .insert(Name::new("Colored smooth Flag"));
 }