+++ src/lib.rs
@@ -58,14 +58,14 @@
 //!         // Add your mesh, material and your custom PBR data   
 //!         ..Default::default()
 //!     }).insert(ClothBuilder::new()
-//!         // Define fixed vertices using an Iterator
-//!         .with_fixed_points(0..9)
+//!         // Define pinned vertices ids using an Iterator
+//!         .with_pinned_vertex_ids(0..9)
 //!         // Define the stick generation mode
 //!         .with_stick_generation(StickGeneration::Quads)
 //!         // Defines the sticks target length option
 //!         .with_stick_length(StickLen::Auto)
 //!         // The cloth will compute flat mesh normals
-//!         .with_flat_normal_computation()
+//!         .with_flat_normals()
 //!         // ...
 //!     );
 //! }
@@ -85,7 +85,8 @@
 //!     .insert_resource(ClothConfig {
 //!         gravity: Vec3::new(0.0, -9.81, 0.0),
 //!         friction: 0.02,
-//!         sticks_computation_depth: 5
+//!         sticks_computation_depth: 5,
+//!         acceleration_smoothing: AccelerationSmoothing::default()
 //!     })
 //!     .add_plugin(ClothPlugin)
 //!     // ... Add your resources and systems
@@ -134,7 +135,12 @@
 //!
 //! - `My mesh falls immediately and infinitely when I add a Cloth component, how to fix it?`
 //!
-//! You probably didn't specify any *fixed points*, meaning there are no vertices anchored to your entity's `GlobalTransform`.
+//! You probably didn't specify any *pinned points*, meaning there are no vertices anchored to your entity's `GlobalTransform`.
+//!
+//! - `My cloth jitters a lot/ suddenly falls down/ has strange sudden behaviour`
+//!
+//! Gravity and winds are bu default smoothed out by the framerate, if the framerate drops suddenly gravity and wind get much stronger.
+//! If your simulation suffers from this you can specify a custom smooth value in `ClothConfig::acceleration_smoothing`.
 //!
 #![forbid(unsafe_code)]
 #![warn(
@@ -179,7 +185,7 @@ pub mod prelude {
     pub use crate::{
         cloth_builder::ClothBuilder,
         cloth_rendering::NormalComputing,
-        config::ClothConfig,
+        config::{AccelerationSmoothing, ClothConfig},
         error::Error,
         mesh::rectangle_mesh,
         stick::{StickGeneration, StickLen},